Moved adding of `flags` context into Scope (#3917)
Using an error_processor to read data from the scope to add to the event is an anti-pattern.  Moving this into `Scope.apply_to_event()`. 

This PR:
- moves code that adds flags to an event from an error processor into the `Scope` class 
- moves `add_feature_flag()` function from `sentry_sdk.integrations.feature_flags` into `sentry_sdk.feature_flags`
