+++ sentry_sdk/feature_flags.py
@@ -1,11 +1,10 @@
-from typing import TYPE_CHECKING
-
 import sentry_sdk
 from sentry_sdk._lru_cache import LRUCache
 
+from typing import TYPE_CHECKING
+
 if TYPE_CHECKING:
-    from typing import TypedDict, Optional
-    from sentry_sdk._types import Event, ExcInfo
+    from typing import TypedDict
 
     FlagData = TypedDict("FlagData", {"flag": str, "result": bool})
 
@@ -33,8 +32,11 @@ class FlagBuffer:
         self.buffer.set(flag, result)
 
 
-def flag_error_processor(event, exc_info):
-    # type: (Event, ExcInfo) -> Optional[Event]
-    scope = sentry_sdk.get_current_scope()
-    event["contexts"]["flags"] = {"values": scope.flags.get()}
-    return event
+def add_feature_flag(flag, result):
+    # type: (str, bool) -> None
+    """
+    Records a flag and its value to be sent on subsequent error events.
+    We recommend you do this on flag evaluations. Flags are buffered per Sentry scope.
+    """
+    flags = sentry_sdk.get_current_scope().flags
+    flags.set(flag, result)
