+++ tests/test_feature_flags.py
@@ -4,15 +4,11 @@ import sys
 import pytest
 
 import sentry_sdk
-from sentry_sdk.integrations.feature_flags import (
-    FeatureFlagsIntegration,
-    add_feature_flag,
-)
+from sentry_sdk.feature_flags import add_feature_flag, FlagBuffer
 
 
 def test_featureflags_integration(sentry_init, capture_events, uninstall_integration):
-    uninstall_integration(FeatureFlagsIntegration.identifier)
-    sentry_init(integrations=[FeatureFlagsIntegration()])
+    sentry_init()
 
     add_feature_flag("hello", False)
     add_feature_flag("world", True)
@@ -34,8 +30,7 @@ def test_featureflags_integration(sentry_init, capture_events, uninstall_integra
 def test_featureflags_integration_threaded(
     sentry_init, capture_events, uninstall_integration
 ):
-    uninstall_integration(FeatureFlagsIntegration.identifier)
-    sentry_init(integrations=[FeatureFlagsIntegration()])
+    sentry_init()
     events = capture_events()
 
     # Capture an eval before we split isolation scopes.
@@ -86,8 +81,7 @@ def test_featureflags_integration_asyncio(
 ):
     asyncio = pytest.importorskip("asyncio")
 
-    uninstall_integration(FeatureFlagsIntegration.identifier)
-    sentry_init(integrations=[FeatureFlagsIntegration()])
+    sentry_init()
     events = capture_events()
 
     # Capture an eval before we split isolation scopes.
@@ -131,3 +125,45 @@ def test_featureflags_integration_asyncio(
             {"flag": "world", "result": False},
         ]
     }
+
+
+def test_flag_tracking():
+    """Assert the ring buffer works."""
+    buffer = FlagBuffer(capacity=3)
+    buffer.set("a", True)
+    flags = buffer.get()
+    assert len(flags) == 1
+    assert flags == [{"flag": "a", "result": True}]
+
+    buffer.set("b", True)
+    flags = buffer.get()
+    assert len(flags) == 2
+    assert flags == [{"flag": "a", "result": True}, {"flag": "b", "result": True}]
+
+    buffer.set("c", True)
+    flags = buffer.get()
+    assert len(flags) == 3
+    assert flags == [
+        {"flag": "a", "result": True},
+        {"flag": "b", "result": True},
+        {"flag": "c", "result": True},
+    ]
+
+    buffer.set("d", False)
+    flags = buffer.get()
+    assert len(flags) == 3
+    assert flags == [
+        {"flag": "b", "result": True},
+        {"flag": "c", "result": True},
+        {"flag": "d", "result": False},
+    ]
+
+    buffer.set("e", False)
+    buffer.set("f", False)
+    flags = buffer.get()
+    assert len(flags) == 3
+    assert flags == [
+        {"flag": "d", "result": False},
+        {"flag": "e", "result": False},
+        {"flag": "f", "result": False},
+    ]
